Added token refresh to tasks/verify-compose-finished.yaml
When checking for the status of one or more compose requests, it is
possible that the token will expire before any of the requests finishes,
as the refresh token is only valid for 15 minutes.

To overcome this limitation, a couple of tasks were added:
 - Compose check loops include less retries by default
 - When the loop of the first image fails, the loop is interrupted
 - When a 401 is detected as reason for failure, token is renewed
 - Loops are retried for (customizable) number of times

Additionally, image builder failures are now properly handled (skipped)
in download and subsequent tasks.