Extend customizations (#13)
* Added code to support all defined customizations

Moved customization fields to the image definition rather than on the
ansible role, so that all defined customizations [0] can be used in the
image definitions.

[0] https://developers.redhat.com/api-catalog/api/image-builder#schema-Customizations

* Increasing retries due to long-running customizations

Some image builds can take over an hour to complete. The new default values of retries
should give enough time for most of the builds to complete.