Make release_id optional (#27)
This change makes usage of `release_id` optional. When defined in yaml,
release_id is used as part of the downloaded file names (eg:
`rhel-8.4-hardened-ib-2024-11-12.qcow2`). If the image is uploaded to
RHOSP, a tag is added with the release_id as well.

When undefined, the name of the file is simply the name of the image
(eg: `rhel-8.4-hardened-ib.qcow2`). Without a release_id to identify the
files, if persistent storage is used, the image is not rotated.