Check payload from Github
Added a playbook to check information from an event in Github, to be
able to decide if further action is needed. If the event is a pull
request merged on the main brach, display release information.
Otherwise, fail.
This prevents images from being requested and rotated for each PR event
that Github sends.