Extend offline customizations
Offline customizations are not the best, but sometimes they are
unavoidable. There's only so much that the Image Builder API can do. For
example, creating a symlink, removing packages, editing the content of a
file or creating a file outside of /etc cannot be done using the API.

This MR extends offline customizations so that multiple different
commands can be passed to `virt-customize` or virt-edit`, which greatly
extends customization options.